#23: Fix dashboard routing, sidebars, and Assign button UX
# Motivation
Customers were getting the Team nav because the sidebar logic compared a backend enum to a string, causing the wrong sidebar and content to render. Team members were also seeing the customer dashboard, and the Profile page was rendering inside the dashboard view instead of its own page. Additionally, admins reported the Assign button in orders was not working — preventing assignment of team members to orders. These issues break role-based navigation and basic admin workflow, so they need to be addressed to restore correct routing, sidebars, profile routing, and assignment functionality.

# Changes
No changes planned.

Project: Cargivo
Draft version: 23

+++ src/frontend/src/components/OrderDetailModal.tsx
@@ -60,7 +60,10 @@ export function OrderDetailModal({
     if (!actor || !newStatus) return;
     setIsUpdating(true);
     try {
-      await actor.updateOrderStatus(order.id, newStatus as OrderStatus);
+      await actor.updateOrderStatus(
+        BigInt(order.id as bigint),
+        newStatus as OrderStatus,
+      );
       toast.success("Order status updated");
       onRefresh?.();
       onClose();
@@ -75,7 +78,7 @@ export function OrderDetailModal({
     if (!actor) return;
     setIsAccepting(true);
     try {
-      await actor.acceptQuote(order.id);
+      await actor.acceptQuote(BigInt(order.id as bigint));
       toast.success("Quote accepted!");
       onRefresh?.();
       onClose();
